Limit build architectures (#315)
As we depend on the elementary PPA for their SDK packages, we should limit Akira to only build on architectures supported by those packages. That means no s390x mainframe builds and no powerpc builds for you! :D
(honestly, it would be hilarious to run elementary os and akira on an S390 mainframe) :D