Use absolute positions in move operation
Snapping requires more robust movement of items. This PR movees
the move_from_event to the SelectedBoundManager.

The SelectedBoundManager stores drag data regarding the start
position of the event and of the item in order to set new positions
based on those instead of relative to an accumulator as before.

This will allow for much nicer snapping