Implement serializatino / deserializatino for Lib2 (#648)
Still a WIP, a lot of checks aren't there, so if something
isn't right it could crash. For the most part it should
be ok. It can be improved with time.

Structure is also not ideal, but not sure what would be the
best way in the future. I mostly want to get the ball rolling,
and I kept serialization and deserialization aspects in
files together to make things easier to follow.

Things could potentially be better with delegates and whatnot,
but meh.